requirements/test.txt
# Test dependencies go here.
-r base.txt

# Python-PostgreSQL Database Adapter
# If using Win for dev, this assumes Unix in test/prod
psycopg2==2.6.1

coverage==4.0.3
django_coverage_plugin==1.3
flake8==2.5.4
django-test-plus==1.0.12
factory_boy==2.7.0

# pytest
pytest-django==2.9.1
pytest-sugar==0.7.1
